+++ PokemonGhost/Normal.h
@@ -5,7 +5,7 @@
 class Normal : public Agent
 {
 public:
-	Normal(int speed, glm::vec2 postition, std::string texture);
+	Normal(float speed, glm::vec2 postition, std::string texture);
 	~Normal();
 
 	void update(Level* level, std::vector<Normal*> normals, std::vector<Ghost*> ghosts) override;
